\name{arsample.unif}
\alias{arsample.unif}
\title{Accept/Reject Algorithm}
\usage{
  arsample.unif(f, M, ...)
}
\arguments{
  \item{f}{target density}

  \item{M}{maximum value for enveloping uniform density}

  \item{...}{additional arguments sent to f}
}
\value{
  x an observation from the target density
}
\description{
  Accept/Reject Algorithm
}
\author{
  Heike Hofmann
}

