\name{as.EA}
\alias{as.EA}
\title{Convert anything into EA class}
\usage{
  as.EA(x)
}
\arguments{
  \item{x}{can be anything}
}
\value{
  x with class "EA"
}
\description{
  Convert anything into EA class
}
\seealso{
  Other EA: \code{\link{EA}}, \code{\link{EA-class}},
  \code{\link{EA.default}}, \code{\link{EA.Q4}},
  \code{\link{EA.SO3}}, \code{\link{id.EA}}
}

