\name{as.Q4}
\alias{as.Q4}
\title{Convert anything into Q4 class}
\usage{
  as.Q4(x)
}
\arguments{
  \item{x}{can be anything}
}
\value{
  x with class "Q4"
}
\description{
  Convert anything into Q4 class
}
\seealso{
  Other Q4: \code{\link{id.Q4}}, \code{\link{Q4}},
  \code{\link{Q4-class}}, \code{\link{Q4.default}},
  \code{\link{Q4.SO3}}
}

