\name{as.SO3}
\alias{as.SO3}
\title{Convert anything into SO3 class}
\usage{
  as.SO3(x)
}
\arguments{
  \item{x}{matrix of rotations, note that no check is
  performed}
}
\value{
  x with class "SO3"
}
\description{
  Convert anything into SO3 class
}
\seealso{
  Other SO3: \code{\link{id.SO3}}, \code{\link{is.Q4}},
  \code{\link{is.SO3}}, \code{\link{SO3}},
  \code{\link{SO3-class}}, \code{\link{SO3.default}},
  \code{\link{SO3.EA}}, \code{\link{SO3.Q4}}
}

