% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa.R
\name{cayley.kappa}
\alias{cayley.kappa}
\title{Circular variance and concentration parameter 
 
 Return the concentration parameter that corresponds to a given circular variance.
 
 The concentration parameter \eqn{\kappa} does not translate across circular distributions.  A commonly used
 measure of spread in circular distributions that does translate
 is the circular variance defined as \eqn{\nu=1-E[\cos(r)]}{\nu=1-E[cos(r)]} where \eqn{E[\cos(r)]}{E[cos(r)]} 
 is the mean resultant length.  See \cite{mardia2000} for more details.
 This function translates the circular variance \eqn{\nu} into the corresponding concentration parameter \eqn{\kappa}
 for the Cayley distribution.
 
 @param nu circular variance
 @return Concentration parameter corresponding to nu.
 @cite mardia2000
 @seealso \code{\link{Cayley}}
 @export
 @examples
 #Find the concentration parameter for circular variances 0.25, 0.5, 0.75
 cayley.kappa(0.25)
 cayley.kappa(0.5)
 cayley.kappa(0.75)}
\usage{
cayley.kappa(nu)
}
\description{
Circular variance and concentration parameter 
 
 Return the concentration parameter that corresponds to a given circular variance.
 
 The concentration parameter \eqn{\kappa} does not translate across circular distributions.  A commonly used
 measure of spread in circular distributions that does translate
 is the circular variance defined as \eqn{\nu=1-E[\cos(r)]}{\nu=1-E[cos(r)]} where \eqn{E[\cos(r)]}{E[cos(r)]} 
 is the mean resultant length.  See \cite{mardia2000} for more details.
 This function translates the circular variance \eqn{\nu} into the corresponding concentration parameter \eqn{\kappa}
 for the Cayley distribution.
 
 @param nu circular variance
 @return Concentration parameter corresponding to nu.
 @cite mardia2000
 @seealso \code{\link{Cayley}}
 @export
 @examples
 #Find the concentration parameter for circular variances 0.25, 0.5, 0.75
 cayley.kappa(0.25)
 cayley.kappa(0.5)
 cayley.kappa(0.75)
}
