% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drill}
\alias{drill}
\title{Drill tests of 8 subjects.}
\format{
A data frame with 720 rows and 8 variables:
\describe{
  \item{Subject}{Subject Id}
  \item{Joint}{Name of the body joint}
  \item{Position}{Position of the body joint}
  \item{Replicate}{Replicate Id}
  \item{Q1}{First component of measured quaternions}
  \item{Q2}{Second component of measured quaternions}
  \item{Q3}{Third component of measured quaternions}
  \item{Q4}{Fourth component of measured quaternions}
}
}
\source{
\url{???}
}
\usage{
drill
}
\description{
A data set containing drill tests for 8 subjects for various body joints.
}
\keyword{datasets}
