\name{eskew}
\alias{eskew}
\title{Directional vector to skew-symmetric Matrix}
\usage{
  eskew(U)
}
\arguments{
  \item{U}{three dimensional vector indicating rotational
  fix-axis}
}
\value{
  skew-symmetric matrix
}
\description{
  Directional vector to skew-symmetric Matrix
}
\author{
  Heike Hofmann
}

