% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_single_locus.r
\name{fitness_single_locus_test}
\alias{fitness_single_locus_test}
\title{test fill a single locus fitness array}
\usage{
fitness_single_locus_test(num_ins = NULL, a_fitloc = NULL, plot = FALSE)
}
\arguments{
\item{num_ins}{number of insecticides}

\item{a_fitloc}{array of single locus fitnesses to fill}

\item{plot}{whether to plot fitness}
}
\value{
fitness values
}
\description{
with hardcoded values
}
\examples{
#defaults
fitness_single_locus()
}
