% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnessSingleLocus.r
\name{fitnessSingleLocus}
\alias{fitnessSingleLocus}
\title{calculate single locus fitness for 2 loci from insecticide exposure
VERSION FROM RESISTANCE NOT USED IN ROTATIONS}
\usage{
fitnessSingleLocus(eff1 = 0.5, eff2 = 0.5, dom1 = 0.5, dom2 = 0.5,
  rr_1 = 0.5, rr_2 = 0.5, cost1 = 0, cost2 = 0, fitSS1 = 1,
  fitSS2 = 1, a_dom = NULL, a_sel = NULL, a_effect = NULL,
  a_cost = NULL, a_fitloc = NULL, plot = FALSE)
}
\arguments{
\item{eff1}{effectiveness1}

\item{eff2}{effectiveness2}

\item{dom1}{dominance1}

\item{dom2}{dominance2}

\item{rr_1}{resistance restoration 1 selection coefficient = resistance restoration * effectiveness}

\item{rr_2}{resistance restoration 2}

\item{cost1}{fitness cost of R in no insecticide}

\item{cost2}{fitness cost of R in no insecticide}

\item{fitSS1}{fitness of SS1 if no insecticide}

\item{fitSS2}{fitness of SS2 if no insecticide}

\item{a_dom}{dominance array}

\item{a_sel}{selection coefficient array}

\item{a_effect}{effectiveness array}

\item{a_cost}{cost array}

\item{a_fitloc}{array of single locus fitnesses to fill}

\item{plot}{whether to plot fitness}
}
\value{
fitness values
}
\description{
can be used in 2 ways
1) by passing arrays a_dom etc. as done from runModel2()
2) by passing separate inputs (e.g. eff1) for a single niche for testing & vis
}
\examples{
fitnessSingleLocus()
fitnessSingleLocus(eff1 = 0.8)
}
