\name{haar}
\alias{haar}
\title{Angle distributions}
\usage{
  haar(x)
}
\arguments{
  \item{x}{vector of angles (between -pi and pi)}

  \item{kappa}{concentration parameter}
}
\description{
  Densities of the most commonly used angle distributions:
  Haar, Fisher, Cayley, and von Mises Distributions
}

