\docType{data}
\name{id.EA}
\alias{id.EA}
\title{Identity in EA space}
\format{Class 'EA'  num [1:3] 0 0 0}
\usage{
  id.EA
}
\description{
  Identity in EA space
}
\seealso{
  Other EA: \code{\link{EA}}, \code{\link{EA-class}},
  \code{\link{EA.Q4}}, \code{\link{EA.SO3}},
  \code{\link{EA.default}}, \code{\link{as.EA}}
}
\keyword{datasets}

