\docType{data}
\name{id.Q4}
\alias{id.Q4}
\title{Identity in Q4 space}
\format{Class 'Q4'  num [1:4] 1 0 0 0}
\usage{
  id.Q4
}
\description{
  Identity in Q4 space
}
\seealso{
  Other Q4: \code{\link{Q4}}, \code{\link{Q4-class}},
  \code{\link{Q4.EA}}, \code{\link{Q4.SO3}},
  \code{\link{Q4.default}}, \code{\link{as.Q4}}
}
\keyword{datasets}

