\docType{data}
\name{id.SO3}
\alias{id.SO3}
\title{Identity in SO(3) space}
\format{SO3 [1:3, 1:3] 1 0 0 0 1 0 0 0 1}
\usage{
  id.SO3
}
\description{
  Identity in SO(3) space
}
\seealso{
  Other SO3: \code{\link{as.SO3}}, \code{\link{is.Q4}},
  \code{\link{is.SO3}}, \code{\link{SO3}},
  \code{\link{SO3-class}}, \code{\link{SO3.default}},
  \code{\link{SO3.Q4}}
}
\keyword{datasets}

