% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insecticide_check.r
\name{insecticide_check}
\alias{insecticide_check}
\title{insecticide_check check if insecticide needs to be changed}
\usage{
insecticide_check(RAF1gen, current_insecticide, rot_interval, rot_criterion,
  gens_this_insecticide, min_rwr_interval)
}
\arguments{
\item{RAF1gen}{single generation array of resistance allele frequencies}

\item{current_insecticide}{id number of current insecticide}

\item{rot_interval}{frequency of rotation (in generations) NB if set to zero mean RwR i.e. rotate when resistant}

\item{rot_criterion}{resistant allele frequency that triggers a RwR change or precludes a insecticide from being rotated in.}

\item{gens_this_insecticide}{num generations this insecticide has been used for}

\item{min_rwr_interval}{minimum rotate-when-resistant interval to stop short switches, only used when rot_interval==0. set to 0 to have no effect.}
}
\value{
TRUE/FALSE whether to change
}
\description{
based on a fixed interval no. generations, or rotate-when-resistant if interval is 0
}
