% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insecticide_switch.r
\name{insecticide_switch}
\alias{insecticide_switch}
\title{insecticide_switch choose a new insecticide to switch to}
\usage{
insecticide_switch(RAF, current_insecticide, n_insecticides, threshold,
  mort_or_freq, gen, min_gens_switch_back, df_ins, df_results, diagnostics,
  eff, dom_sel, rr)
}
\arguments{
\item{RAF}{array of resistance allele frequencies}

\item{current_insecticide}{id number of current insecticide}

\item{n_insecticides}{number of insecticides (and hence loci)}

\item{threshold}{trigger for change of insecticide, either resistance frequency or mortality dependent on mort_or_freq, also precludes switch to an insecticide.}

\item{mort_or_freq}{whether threshold for insecticide change is mortality 'mort' or resistance frequency 'freq'}

\item{gen}{generation number}

\item{min_gens_switch_back}{minimum num gens before can switch back to an insecticide}

\item{df_ins}{number of generations since each insecticide used}

\item{df_results}{results of sim so far}

\item{diagnostics}{whether to output running info}

\item{eff}{effectiveness, for all insecticides or individually}

\item{dom_sel}{dominance of selection, for all insecticides or individually}

\item{rr}{resistance restoration, for all insecticides or individually}
}
\value{
integer : insecticide number to switch to, if 0 none found
}
\description{
only gets called if insecticide_check() has already identified that the previous insecticide needed to be changed.   
Checks that the new insecticide is below resistance or survival thresholds.
}
