% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration_test.r
\name{migration_test}
\alias{migration_test}
\title{migration_test to testimplementation of migration between an implementation area and refugia}
\usage{
migration_test(migration = 0.5, coverage = 0.6, startfreqs = c(0.2, 0.8),
  max_generations = 10)
}
\arguments{
\item{coverage}{}
}
\description{
migration_test to testimplementation of migration between an implementation area and refugia
}
