% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nickel}
\alias{nickel}
\title{Nickel data set.}
\format{
A data frame with 48,286 rows and 13 variables:
\describe{
  \item{xpos}{X-axis coordinate}
  \item{ypos}{Y-axis coordinate}
  \item{location}{Location Id}
  \item{rep}{Replicate Id}
  \item{V1}{First entry of measured rotation matrices}
  \item{V2}{Second entry of measured rotation matrices}
  \item{V3}{Third entry of measured rotation matrices}
  \item{V4}{Fourth entry of measured rotation matrices}
  \item{V5}{Fifth entry of measured rotation matrices}
  \item{V6}{Sixth entry of measured rotation matrices}
  \item{V7}{Seventh entry of measured rotation matrices}
  \item{V8}{Eighth entry of measured rotation matrices}
  \item{V9}{Nineth entry of measured rotation matrices}
}
}
\source{
\url{???}
}
\usage{
nickel
}
\description{
A data set containing nickel data.
}
\keyword{datasets}
