\docType{package}
\name{package-rotations}
\alias{package-rotations}
\alias{package-rotations-package}
\title{A package for working with rotaional data.}
\description{
  This package implements tools for working with rotational
  data: it allows simulation from the most commonly used
  distributions on the sphere, it estimates differen mean
  and median type estimators for the main direction of a
  number of rotations, it provides (bootstrap) confidence
  regions for the estimates, and it allows to visualize
  rotational data.
}

