\name{Phi}
\alias{Phi}
\title{Skew-symmetric matrix corresponding to rotation R}
\usage{
  Phi(U)
}
\arguments{
  \item{U}{three-dimensional vector of real numbers
  describing the fix axis of rotation R.}
}
\description{
  The space of skew-symmetric matrices, i.e. all square
  matrices A with \eqn{A^t = -A} is the tangent space of
  the special orthogonal group \eqn{SO(3)}, that describes
  all three-dimensional rotations.
}

