% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot_migrate.r
\name{rot_migrate}
\alias{rot_migrate}
\title{migration between intervention and refugia}
\usage{
rot_migrate(RAF1gen, migration = 0.8, coverage = 0.6, verbose = FALSE)
}
\arguments{
\item{RAF}{array of resistance allele frequencies}

\item{migrate_intervention}{migration from intervention}

\item{migrate_refugia}{migration from refugia}
}
\value{
array of resistance freqs in 1 gen following migration
}
\description{
migration between intervention and refugia
}
\examples{
#frequencies the same for all insecticides
RAF <- set_start_freqs(n=3, freqs=0.001)
RAF[,,,1] # to view generation 1
#frequencies different for each insecticide
RAF <- set_start_freqs(freqs=c(0.1,0.01,0.001))
RAF[,,,1] # to view generation 1
#testing mutliplying frequency arrays
RAF <- set_start_freqs(max_generations = 1, freqs=c(0.1,0.01,0.001))
RAF2 <- set_start_freqs(max_generations = 1, freqs=c(1,2,3))
RAF*RAF2


}
