% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rot.r
\name{run_rot}
\alias{run_rot}
\title{run_rot to run simulation of the effect of rotations on the spread of resistance}
\usage{
run_rot(max_generations = 200, n_insecticides = 4, start_freqs = 0.001,
  rotation_interval = 0, rotation_criterion = 0.5,
  migration_rate_intervention = 0.01, coverage = 0.8, plot = TRUE,
  start_insecticide = 1, diagnostics = FALSE, hardcode_fitness = TRUE,
  same_insecticides = TRUE, hardcode_exposure = TRUE, expo_hi = 0.8,
  expo_lo = 0, male_expo_prop = 1)
}
\arguments{
\item{max_generations}{maximum number of mosquito generations to run the simulation}

\item{n_insecticides}{number of insecticides (and hence loci), current max is 5}

\item{start_freqs}{starting frequencies of resistance either one per insecticide or same for all}

\item{rotation_interval}{frequency of rotation (in generations) NB if set to zero mean RwR i.e. rotate when resistant}

\item{rotation_criterion}{resistant allele frequency that triggers a RwR change or precludes a insecticide from being rotated in.}

\item{migration_rate_intervention}{migration rate into and out-of the treated area. It is the proportion of the treated population that migrates. We assume that immigration=emigration.}

\item{coverage}{proportion of mosquitoes that are covered by the intervention (and 1-C is the proportion of the population in the untreated refugia).}

\item{plot}{whether to plot results}

\item{start_insecticide}{which insecticide to start with}

\item{diagnostics}{whether to output running info}

\item{same_insecticides}{only used with hardcode_fitness, whether to just set fitnesses for all insecticides the same}
}
\value{
dataframe of results
}
\description{
can run any number of insecticides/loci 
but at present, input will only allow a maximumum of 5
}
\examples{
run_rot(rotation_interval=100)
dfr <- run_rot(rotation_interval=50, max_generations = 300)
dfr <- run_rot(rotation_interval=0, max_generations = 300)
dfr <- run_rot(rotation_interval=0, max_generations = 300, hardcode_fitness = TRUE, same_insecticides =TRUE, migration_rate_intervention=0.01)

}
