% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_exposure_rot.r
\name{set_exposure_rot}
\alias{set_exposure_rot}
\title{set exposure to insecticides for rotations to a flexible number of insecticides}
\usage{
set_exposure_rot(num_ins = NULL, expo_hi = 0.8, expo_lo = 0,
  male_expo_prop = 1, plot = FALSE)
}
\arguments{
\item{num_ins}{number of insecticides, optional can just be specified by number of items in vector expo}

\item{male_expo_prop}{proportion tht males are exposed relative to f, default 1, likely to be <1}

\item{plot}{whether to plot exposure}

\item{expo}{exposure to the insecticides}
}
\value{
array of exposure values for the different insecticides
}
\description{
TODO could this be made generic to the resistance code too ? 
I suspect not. They do things quite differently.
This just needs 1 insecticide at a time, so is :
array_named(insecticide=1:num_ins, sex=c('m','f'), exposure=c('no','lo','hi')
In resistance it is :
array_named( sex=c('m','f'), niche1=c('0','a','A'), niche2=c('0','b','B') )
}
\details{
fills an array of exposure values
}
\examples{
a_exp <- set_exposure_rot( expo_hi=0.9 )
a_exp <- set_exposure_rot( expo_hi=c(0.5,0.9), male_expo_prop = 0.9)

#allowing array to be viewed differently
as.data.frame(a_exp)

}
