% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_exposure_rot.r
\name{set_exposure_rot_test}
\alias{set_exposure_rot_test}
\title{set_exposure_rot_test allows exposures to be hardcoded}
\usage{
set_exposure_rot_test(num_ins = NULL, plot = FALSE)
}
\arguments{
\item{num_ins}{number of insecticides, optional can just be specified by number of items in vector expo}

\item{plot}{whether to plot exposure}
}
\value{
array of exposure values for the different insecticides
}
\description{
set exposure to insecticides for rotations to a flexible number of insecticides
fills an array of exposure values
}
\examples{
a_exp <- set_exposure_rot_test( )

#allowing array to be viewed differently
as.data.frame(a_exp)

}
