% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_start_freqs.r
\name{set_start_freqs}
\alias{set_start_freqs}
\title{set_start_freqs allows starting allele frequencies to be set}
\usage{
set_start_freqs(n_insecticides = NULL, max_generations = 200,
  freqs = 0.001)
}
\arguments{
\item{n_insecticides}{number of insecticides, optional can just be specified by number of items in vector expo}

\item{freqs}{starting allele frequencies either one per insecticide or same for all
in this version same for m,f,intervention & refuge}
}
\value{
array of allele frequencies with just first record filled
}
\description{
set_start_freqs allows starting allele frequencies to be set
}
\examples{
#frequencies the same for all insecticides
RAF <- set_start_freqs(n=3, freqs=0.001)
RAF[,,,1] # to view generation 1
#frequencies different for each insecticide
RAF <- set_start_freqs(n=3, freqs=c(0.1,0.01,0.001))
RAF[,,,1] # to view generation 1
#testing mutliplying frequency arrays
RAF <- set_start_freqs(n=3, max_generations = 1, freqs=c(0.1,0.01,0.001))
RAF2 <- set_start_freqs(n=3, max_generations = 1, freqs=c(1,2,3))
RAF*RAF2

#allowing array to be viewed differently
as.data.frame(RAF)

}
