% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_start_freqs.r
\name{set_start_freqs_test}
\alias{set_start_freqs_test}
\title{set_start_freqs_test allows starting allele frequencies to be hardcoded}
\usage{
set_start_freqs_test(n_insecticides = NULL, max_gen = 200,
  plot = FALSE)
}
\arguments{
\item{n_insecticides}{number of insecticides, optional can just be specified by number of items in vector expo}

\item{max_gen}{maximum number of mosquito generations to run the simulation}

\item{plot}{whether to plot exposure}
}
\value{
array of allele frequencies with just first record filled
}
\description{
set_start_freqs_test allows starting allele frequencies to be hardcoded
}
\examples{
RAF <- set_start_freqs_test( )

#allowing array to be viewed differently
as.data.frame(RAF)

}
