\name{vecNorm}
\alias{vecNorm}
\title{Turn a vector into a matrix and compute the the norm between x and S}
\usage{
  vecNorm(x, S, ...)
}
\arguments{
  \item{x}{numeric vector}

  \item{S}{numeric vector or matrix}

  \item{...}{additional arguments passed to
  \code{\link{norm}} function}
}
\value{
  the norm of x-S
}
\description{
  Turn a vector into a matrix and compute the the norm
  between x and S
}

