% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Angular-distributions}
\alias{Angular-distributions}
\title{Angular distributions}
\description{
Density, distribution function and random variate generation for symmetric probability distributions in the rotations package.
}
\details{
The functions for the density function and random variate generation are named in the usual form dxxxx, pxxxx and rxxxx,
respectively.
\itemize{
	\item See \code{\link{Cayley}} for the Cayley distribution.
	\item See \code{\link{Fisher}} for the matrix Fisher distribution.
	\item See \code{\link{Haar}} for the uniform distribution on the circle.
	\item See \code{\link{Maxwell}} for the Maxwell-Boltzmann distribution on the circle.
	\item See \code{\link{Mises}} for the von Mises-Fisher distribution.
}
}
