\name{Angular-distributions}
\alias{Angular-distributions}
\title{Angular distributions in the rotations packages}
\description{
  Density and random variate generation for symmetric
  probability distributions in the rotations package
}
\details{
  The functions for the density function and random variate
  generation are named in the usual form dxxxx and rxxxx
  respectively.  See \code{\link{dcayley}} for the Cayley
  distribution.  See \code{\link{dvmises}} for the von
  Mises-Fisher distribution.  See \code{\link{dhaar}} for
  the uniform distribution on the circle.  See
  \code{\link{dfisher}} for the matrix Fisher distribution.
}

