\name{arsample.unif}
\alias{arsample.unif}
\title{Accept reject ratio of uniforms?  Written by Dr. Hofmann}
\usage{
  arsample.unif(f, M, ...)
}
\arguments{
  \item{f}{target density}

  \item{M}{maximum value for one of the uniforms}

  \item{...}{additional arguments sent to f}
}
\value{
  x an observation from the target density
}
\description{
  Accept reject ratio of uniforms?  Written by Dr. Hofmann
}
\author{
  Heike Hofmann
}

