\name{dhaar}
\alias{dhaar}
\title{Haar Measure}
\usage{
  dhaar(r)
}
\arguments{
  \item{r}{Where the density is being evaluated}
}
\value{
  the probability density evaluated at r
}
\description{
  The uniform distribution on the sphere is also know as
  the Haar measure and has the density function
  \deqn{C_U(r)=\frac{1-cos(r)}{2\pi}}
}
\seealso{
  \code{\link{rhaar}},
  \code{\link{dfisher}},\code{\link{dvmises}},\code{\link{dcayley}}
}

