\name{discord}
\alias{discord}
\title{Measure of Discord}
\usage{
  discord(x, type, t, ...)
}
\arguments{
  \item{x}{The sample of random rotations}

  \item{t}{If test blocs then the bloc size}

  \item{type}{To specify if "intrinsic" or "extrinsic"
  approach should be used to compute the statistic}
}
\value{
  The list of Hi statistics
}
\description{
  This function computes a measure of discord for a sample
  of random rotations. The larger the statistic value the
  less likely it is the corresponding observation was
  generated by the same mechanism the rest of the data as
  generated by. It can be used to test for outliers in
  SO(3) by comparing it to an F distribution with 3,3(n-2)
  df for the Cayley or matrix Fisher distributions or to an
  F distribution with 1,n-2 df for the von Mises Fisher
  distribution.
}
\examples{
#Compute the measures of discord for a sample from the Cayley distribution
Rs <- ruars(20,rcayley,kappa=50)
Hi <- discord(Rs)

#Compare to the theoretical F distribution
OrdHi <- sort(Hi)

plot(ecdf(OrdHi))
lines(OrdHi,pf(OrdHi,3,3*(length(OrdHi)-2)))
}

