% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insecticide_switch.r
\name{insecticide_switch}
\alias{insecticide_switch}
\title{insecticide_switch choose a new insecticide to switch to}
\usage{
insecticide_switch(RAF, current_insecticide, n_insecticides, rot_criterion, gen,
  min_gens_switch_back, df_ins, df_results, diagnostics)
}
\arguments{
\item{RAF}{array of resistance allele frequencies}

\item{current_insecticide}{id number of current insecticide}

\item{n_insecticides}{number of insecticides (and hence loci)}

\item{rot_criterion}{resistant allele frequency that triggers a RwR change or precludes a insecticide from being rotated in.}

\item{gen}{generation number}

\item{min_gens_switch_back}{minimum num gens before can switch back to an insecticide}

\item{df_ins}{number of generations since each insecticide used}

\item{df_results}{results of sim so far}

\item{diagnostics}{whether to output running info}
}
\value{
integer : insecticide number to switch to, if 0 none found
}
\description{
insecticide_switch choose a new insecticide to switch to
}
