% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration_test.r
\name{migration_test}
\alias{migration_test}
\title{migration_test to test implementation of migration between treated area and refugia}
\usage{
migration_test(migration = 0.5, coverage = 0.6, startfreqs = c(0.2,
  0.8), max_gen = 10)
}
\arguments{
\item{migration}{migration between 0-1}

\item{coverage}{proportion of popn in treated area}

\item{startfreqs}{vector of 2 starting resistance frequencies, first in treated area, 2nd in refuge}

\item{max_gen}{maximum number of mosquito generations to run the simulation}
}
\value{
array of resistance frequencies in active area and refuge
}
\description{
migration_test to test implementation of migration between treated area and refugia
}
