% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminary.R
\name{project.SO3}
\alias{project.SO3}
\title{Projection into SO(3)}
\usage{
project.SO3(M)
}
\arguments{
\item{M}{\eqn{3\times 3}{3-by-3} matrix to project into \eqn{SO(3)}.}
}
\value{
Projection of \eqn{\bm M}{M} into \eqn{SO(3)}.
}
\description{
Project an arbitrary \eqn{3\times 3}{3-by-3} matrix into \eqn{SO(3)}.
}
\details{
This function uses the process detailed in Section 3.1 of \cite{moakher02} to project an arbitrary \eqn{3\times 3}{3-by-3} matrix into \eqn{SO(3)}.
More specifically it finds the closest orthogonal 3-by-3 matrix with determinant one to the provided matrix.
}
\examples{
#Project an arbitrary 3x3 matrix into SO(3)
M<-matrix(rnorm(9), 3, 3)
project.SO3(M)

#Project a sample arithmetic mean into SO(3), same as 'mean'
Rs <- ruars(20, rcayley)
Rbar <- colSums(Rs)/nrow(Rs)
project.SO3(Rbar)              #The following is equivalent
mean(Rs)
}
\seealso{
\code{\link{mean.SO3}}, \code{\link{median.SO3}}
}
