% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot_migrate.r
\name{rot_migrate}
\alias{rot_migrate}
\title{migration between intervention and refugia}
\usage{
rot_migrate(RAF1gen, migration = 0.8, coverage = 0.6, verbose = FALSE)
}
\arguments{
\item{RAF1gen}{single generation array of resistance allele frequencies}

\item{migration}{migration rate between treated & untreated areas 0-1. We assume that immigration=emigration.}

\item{coverage}{proportion of mosquitoes that are covered by the intervention (and 1-C is the proportion of the population in the untreated refugia).}

\item{verbose}{whether to output before and after migration}
}
\value{
array of resistance freqs in 1 gen following migration
}
\description{
migration between intervention and refugia
}
\examples{
#frequencies the same for all insecticides
RAF <- set_start_freqs(n=3, freqs=0.001)
RAF[,,,1] # to view generation 1
#frequencies different for each insecticide
RAF <- set_start_freqs(freqs=c(0.1,0.01,0.001))
RAF[,,,1] # to view generation 1
#testing mutliplying frequency arrays
RAF <- set_start_freqs(max_gen = 1, freqs=c(0.1,0.01,0.001))
RAF2 <- set_start_freqs(max_gen = 1, freqs=c(1,2,3))
RAF*RAF2


}
