% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot_migrate.r
\name{rot_migrate}
\alias{rot_migrate}
\title{migration between intervention and refugia}
\usage{
rot_migrate(RAF1gen, migration = 0.8, coverage = 0.6,
  verbose = FALSE)
}
\arguments{
\item{RAF1gen}{single generation array of resistance allele frequencies}

\item{migration}{migration rate between treated & untreated areas 0-1. We assume that immigration=emigration.}

\item{coverage}{proportion of mosquitoes that are covered by the intervention (and 1-C is the proportion of the population in the untreated refugia).}

\item{verbose}{whether to output before and after migration}
}
\value{
array of resistance freqs in 1 gen following migration
}
\description{
migration between intervention and refugia
}
\examples{

#frequencies different for each insecticide
RAF <- set_start_freqs(freqs=c(0.1,0.01,0.001))

simplest example, freqs 1 in intervention, 0 in refugia
RAF <- set_start_freqs(n=3, freqs=1)
RAF1gen <- RAF[,,,1] # to view generation 1
#set freqs in refugia to 0
RAF1gen[,,'refugia']  <- 0
#testing migration
RAF2 <- rot_migrate(RAF1gen, migration=0.5, coverage=0.5, verbose=TRUE)

#max migration leads to random mixing, irrespective of coverage, intervention and refugia end with same freq
RAF2 <- rot_migrate(RAF1gen, migration=1, coverage=0.5, verbose=TRUE)
RAF2 <- rot_migrate(RAF1gen, migration=1, coverage=0.2, verbose=TRUE)


}
