% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot_plots.r
\name{rot_plot_resistance}
\alias{rot_plot_resistance}
\title{rot_plot_resistance plot rotation simulation resistance results}
\usage{
rot_plot_resistance(df_res2, plot_refuge = TRUE, threshold = 0.5,
  mort_or_freq = "mort", logy = TRUE, add_gens_under50 = TRUE,
  df_resanother = NULL, lwd = 1.5, title = NULL, plot = TRUE,
  plot_mort = TRUE)
}
\arguments{
\item{df_res2}{dataframe of resistance results from run_rot()}

\item{plot_refuge}{whether to plot refuge as well as intervention}

\item{threshold}{trigger for change of insecticide, either resistance frequency or mortality dependent on mort_or_freq, also precludes switch to an insecticide.}

\item{mort_or_freq}{whether threshold for insecticide change is mortality 'mort' or resistance frequency 'freq'}

\item{logy}{whether to use log scale for y axis}

\item{add_gens_under50}{whether to add a label of num generations under 50percent resistance}

\item{df_resanother}{exploratory option to plot results of another scenario on the same graph}

\item{lwd}{line thickness for resistance curves}

\item{title}{optional title for plot, NULL for no title}

\item{plot}{whether to plot results}

\item{plot_mort}{whether to add mortality to plots}
}
\value{
ggplot object
}
\description{
rot_plot_resistance plot rotation simulation resistance results
}
\examples{
df_res2 <- run_rot()
rot_plot_resistance(df_res2)


}
