% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot_plots.r
\name{rot_plot_resistance}
\alias{rot_plot_resistance}
\title{rot_plot_resistance plot rotation simulation resistance results}
\usage{
rot_plot_resistance(df_res2, plot_refuge = TRUE, logy = FALSE,
  add_gens_under50 = TRUE)
}
\arguments{
\item{df_res2}{dataframe of resistance results from run_rot()}

\item{logy}{whether to use log scale for y axis}

\item{add_gens_under50}{whether to add a label of num generations under 50% resistance}

\item{refuge}{whether to plot refuge as well as intervention}
}
\value{
ggplot object
}
\description{
rot_plot_resistance plot rotation simulation resistance results
}
\examples{
df_res2 <- run_rot()
rot_plot_resistance(df_res2)

}
