% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot_plots.r
\name{rot_plot_resistance}
\alias{rot_plot_resistance}
\title{rot_plot_resistance plot rotation simulation resistance results}
\usage{
rot_plot_resistance(df_res2)
}
\arguments{
\item{df_res2}{dataframe of resistance results from run_rot()}
}
\description{
rot_plot_resistance plot rotation simulation resistance results
}
