% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{rotations}
\alias{rotations}
\title{A package for working with rotation data.}
\description{
This package implements tools for working with rotational data: it allows simulation from the 
most commonly used  distributions on \eqn{SO(3)}, it includes methods for different mean and median type 
estimators for the central orientation of a sample, it provides confidence regions 
for those estimates and it includes a novel visualization technique for rotation data.
}
