% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.r
\docType{package}
\name{rotations}
\alias{rotations}
\title{Simulating the Evolution of Insecticide Resistance under Rotations}
\description{
Simulating the Evolution of Insecticide Resistance under Rotations
}
