% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rot.r
\name{run_rot}
\alias{run_rot}
\title{run_rot to run simulation of the effect of rotations on the spread of resistance}
\usage{
run_rot(max_gen = 200, n_insecticides = 4, start_freqs = 0.01,
  rot_interval = 10, rot_criterion = 0.5, migration = 0.01,
  coverage = 0.8, start_insecticide = 1, expo_hi = 0.8, expo_lo = 0,
  male_expo_prop = 1, eff = 0.8, dom_sel = 0.5, dom_cos = 0.5,
  rr = 0.5, cost = 0.1, fitSS = 1, min_rwr_interval = 1,
  no_r_below_start = TRUE, no_r_below_mut = FALSE, exit_rot = TRUE,
  min_gens_switch_back = 10, plot = TRUE, diagnostics = FALSE,
  logy = TRUE, add_gens_under50 = FALSE)
}
\arguments{
\item{max_gen}{maximum number of mosquito generations to run the simulation}

\item{n_insecticides}{number of insecticides (and hence loci)}

\item{start_freqs}{starting frequencies of resistance either one per insecticide or same for all}

\item{rot_interval}{frequency of rotation (in generations) NB if set to zero mean RwR i.e. rotate when resistant}

\item{rot_criterion}{resistant allele frequency that triggers a RwR change or precludes a insecticide from being rotated in.}

\item{migration}{migration rate between treated & untreated areas 0-1. We assume that immigration=emigration.}

\item{coverage}{proportion of mosquitoes that are covered by the intervention (and 1-C is the proportion of the population in the untreated refugia).}

\item{start_insecticide}{which insecticide to start with}

\item{expo_hi}{exposure to insecticide in hi niche, either single or vector of 1 per insecticide}

\item{expo_lo}{exposure to insecticide in lo niche, either single or vector of 1 per insecticide}

\item{male_expo_prop}{proportion tht males are exposed relative to f, default 1, likely to be <1 (could possibly be a vector per insecticide)}

\item{eff}{effectiveness, for all insecticides or individually}

\item{dom_sel}{dominance of selection, for all insecticides or individually}

\item{dom_cos}{dominance of cost, for all insecticides or individually}

\item{rr}{resistance restoration, for all insecticides or individually}

\item{cost}{fitness cost of RR in no insecticide, for all insecticides or individually}

\item{fitSS}{fitness of SS if no insecticide, for all insecticides or individually}

\item{min_rwr_interval}{minimum rotate-when-resistant interval to stop short switches, only used when rot_interval==0. set to 0 to have no effect.}

\item{no_r_below_start}{to stop resistance frequencies going below starting values TRUE or FALSE}

\item{no_r_below_mut}{to stop resistance frequencies going below mutation-selection balance TRUE or FALSE}

\item{exit_rot}{whether to exit rotation interval if rot_criterion is reached}

\item{min_gens_switch_back}{minimum num gens before can switch back to an insecticide}

\item{plot}{whether to plot results}

\item{diagnostics}{whether to output running info}

\item{logy}{whether to use log scale for y axis}

\item{add_gens_under50}{whether to add a label of num generations under 50 pcent resistance}
}
\value{
dataframe of results
}
\description{
can run any number of insecticides/loci
}
\examples{
run_rot(rot_interval=100)
dfr <- run_rot(rot_interval=50, max_gen = 300)
dfr <- run_rot(rot_interval=0, max_gen = 300)
dfr <- run_rot(rot_interval=0, max_gen = 300, migration=0.01)
#running for insecticides with different inputs
dfr <- run_rot(n_insecticides=3, eff=c(1,0.6,0.4), rot_interval=0)

}
