% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rot.r
\name{run_rot}
\alias{run_rot}
\title{run_rot to run simulation of the effect of rotations on the spread of resistance}
\usage{
run_rot(max_generations = 500, n_insecticides = 4, rotation_interval = 0,
  rotation_criterion = 0.5, migration_rate_intervention = 0.01,
  coverage = 0.8, plot = TRUE, diagnostics = FALSE)
}
\arguments{
\item{max_generations}{maximum number of mosquito generations to run the simulation}

\item{n_insecticides}{MAX is 5<<<<the number of insecticides (and hence loci) in the simuation MAX IS 5<<<}

\item{rotation_interval}{frequency of rotation (in generations) NB if set to zero mean RwR i.e. rotate when resistant}

\item{rotation_criterion}{resistant allele frequency that triggers a RwR change or precludes a insecticide from being rotated in.}

\item{migration_rate_intervention}{migration rate into and out-of the treated area. It is the proportion of the treated population that migrates. We assume that immigration=emigration.}

\item{coverage}{proportion of mosquitoes that are covered by the intervention (and 1-C is the proportion of the population in the untreated refugia).}

\item{plot}{whether to plot results}

\item{diagnostics}{whether to output running info}
}
\value{
dataframe of results
}
\description{
can run any number of insecticides/loci 
but at present, input will only allow a maximumum of 5
}
\examples{
run_rot(rotation_interval=100)
df_res2 <- run_rotation()

}
